prepare & minor bugfixes
